Added ActionTimer and entities spawn refactored

+++ BioHeal/Assets/Scripts/JsonEntities/AllyJson.cs
@@ -2,5 +2,5 @@ public class AllyJson : UnitJson
 {
     public int price;
     public int initialC;
-    public int timeToSpawn;
+    public float timeToSpawn;
 }
